Check Stability Bug Fix: M -> stable_matching
Check stability is only called when we do have some stable_matching. Clearly we don't want to check against the provisional assignment graph for two reasons:
a) We damage it with pop() now when we're saving (though clearly this can be avoided if need be);
b) This variable is named as such because it contains the thing to be checked.