Bug Fix - Tied Super Stability Checking
We don't want to have a case where we compare a hospital to itself. This was being handled by removing this hospital from the list of hospitals that a given resident could form blocking pairs with.
This approach affected the original lists. Instead of trying to be clever about it, I've implemented a simple equality check.